requirements.txt
requests>=2.31.0
python-gitlab>=4.4.0
gitpython>=3.1.40
openai>=1.12.0
pyyaml>=6.0.1
jinja2>=3.1.3
APScheduler>=3.10.0
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0
python-dateutil>=2.8.0
openpyxl>=3.1.2
